apps/cv-langserve/pyproject.toml
[tool.poetry]
name = "cv-langserve"
version = "0.1.0"
description = ""
authors = ["Dave Hudson <dave@applification.net>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = ">=0.23.2,<0.24.0"
langserve = {extras = ["server"], version = ">=0.0.39"}
pydantic = "<2"
langchain-community = ">=0.0.13,<0.1"
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"
langchain-openai = "^0.0.2.post1"
tiktoken = "^0.5.2"
openai = "^1.8.0"
pinecone-client = "3.0.0"
ruff = "^0.1.13"
faiss-cpu = "^1.7.4"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"


[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
